backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==10.1.0
pptree==3.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-rapidjson==1.12
python-ulid==2.2.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
